Refactor MVI extensions
