+++ enrollmentproof/src/main/kotlin/com/gdavidpb/tuindice/enrollmentproof/presentation/contract/Enrollment.kt
@@ -1,8 +1,8 @@
 package com.gdavidpb.tuindice.enrollmentproof.presentation.contract
 
-import com.gdavidpb.tuindice.base.presentation.reducer.ViewAction
-import com.gdavidpb.tuindice.base.presentation.reducer.ViewEvent
-import com.gdavidpb.tuindice.base.presentation.reducer.ViewState
+import com.gdavidpb.tuindice.base.utils.extension.ViewAction
+import com.gdavidpb.tuindice.base.utils.extension.ViewEvent
+import com.gdavidpb.tuindice.base.utils.extension.ViewState
 
 object Enrollment {
 	sealed class State : ViewState {
