+++ enrollmentproof/src/main/kotlin/com/gdavidpb/tuindice/enrollmentproof/presentation/reducer/EnrollmentProofReducer.kt
@@ -2,7 +2,7 @@ package com.gdavidpb.tuindice.enrollmentproof.presentation.reducer
 
 import com.gdavidpb.tuindice.base.domain.usecase.base.UseCaseState
 import com.gdavidpb.tuindice.base.presentation.reducer.BaseReducer
-import com.gdavidpb.tuindice.base.presentation.reducer.ViewOutput
+import com.gdavidpb.tuindice.base.utils.extension.ViewOutput
 import com.gdavidpb.tuindice.enrollmentproof.domain.usecase.error.FetchEnrollmentProofError
 import com.gdavidpb.tuindice.enrollmentproof.presentation.contract.Enrollment
 import kotlinx.coroutines.flow.Flow
