+++ evaluations/src/main/kotlin/com/gdavidpb/tuindice/evaluations/presentation/reducer/AddEvaluationReducer.kt
@@ -2,7 +2,7 @@ package com.gdavidpb.tuindice.evaluations.presentation.reducer
 
 import com.gdavidpb.tuindice.base.domain.usecase.base.UseCaseState
 import com.gdavidpb.tuindice.base.presentation.reducer.BaseReducer
-import com.gdavidpb.tuindice.base.presentation.reducer.ViewOutput
+import com.gdavidpb.tuindice.base.utils.extension.ViewOutput
 import com.gdavidpb.tuindice.evaluations.domain.usecase.error.EvaluationError
 import com.gdavidpb.tuindice.evaluations.presentation.contract.Evaluations
 import kotlinx.coroutines.flow.Flow
