+++ evaluations/src/main/kotlin/com/gdavidpb/tuindice/evaluations/presentation/reducer/EvaluationReducer.kt
@@ -3,7 +3,7 @@ package com.gdavidpb.tuindice.evaluations.presentation.reducer
 import com.gdavidpb.tuindice.base.domain.model.Evaluation
 import com.gdavidpb.tuindice.base.domain.usecase.base.UseCaseState
 import com.gdavidpb.tuindice.base.presentation.reducer.BaseReducer
-import com.gdavidpb.tuindice.base.presentation.reducer.ViewOutput
+import com.gdavidpb.tuindice.base.utils.extension.ViewOutput
 import com.gdavidpb.tuindice.base.utils.ResourceResolver
 import com.gdavidpb.tuindice.evaluations.presentation.contract.Evaluations
 import com.gdavidpb.tuindice.evaluations.presentation.mapper.toEvaluationViewState
