+++ login/src/main/kotlin/com/gdavidpb/tuindice/login/presentation/reducer/UpdatePasswordReducer.kt
@@ -2,7 +2,7 @@ package com.gdavidpb.tuindice.login.presentation.reducer
 
 import com.gdavidpb.tuindice.base.domain.usecase.base.UseCaseState
 import com.gdavidpb.tuindice.base.presentation.reducer.BaseReducer
-import com.gdavidpb.tuindice.base.presentation.reducer.ViewOutput
+import com.gdavidpb.tuindice.base.utils.extension.ViewOutput
 import com.gdavidpb.tuindice.login.domain.usecase.error.SignInError
 import com.gdavidpb.tuindice.login.presentation.contract.UpdatePassword
 import kotlinx.coroutines.flow.Flow
