+++ login/src/main/kotlin/com/gdavidpb/tuindice/login/presentation/viewmodel/SignInViewModel.kt
@@ -1,6 +1,6 @@
 package com.gdavidpb.tuindice.login.presentation.viewmodel
 
-import com.gdavidpb.tuindice.base.presentation.reducer.collect
+import com.gdavidpb.tuindice.base.utils.extension.collect
 import com.gdavidpb.tuindice.base.presentation.viewmodel.BaseViewModel
 import com.gdavidpb.tuindice.login.domain.usecase.SignInUseCase
 import com.gdavidpb.tuindice.login.domain.usecase.param.SignInParams
