+++ record/src/main/kotlin/com/gdavidpb/tuindice/record/presentation/reducer/RecordReducer.kt
@@ -3,7 +3,7 @@ package com.gdavidpb.tuindice.record.presentation.reducer
 import com.gdavidpb.tuindice.base.domain.model.quarter.Quarter
 import com.gdavidpb.tuindice.base.domain.usecase.base.UseCaseState
 import com.gdavidpb.tuindice.base.presentation.reducer.BaseReducer
-import com.gdavidpb.tuindice.base.presentation.reducer.ViewOutput
+import com.gdavidpb.tuindice.base.utils.extension.ViewOutput
 import com.gdavidpb.tuindice.base.utils.ResourceResolver
 import com.gdavidpb.tuindice.record.domain.usecase.error.GetQuartersError
 import com.gdavidpb.tuindice.record.presentation.contract.Record
