+++ record/src/main/kotlin/com/gdavidpb/tuindice/record/presentation/viewmodel/RecordViewModel.kt
@@ -1,6 +1,6 @@
 package com.gdavidpb.tuindice.record.presentation.viewmodel
 
-import com.gdavidpb.tuindice.base.presentation.reducer.collect
+import com.gdavidpb.tuindice.base.utils.extension.collect
 import com.gdavidpb.tuindice.base.presentation.viewmodel.BaseViewModel
 import com.gdavidpb.tuindice.record.domain.usecase.GetQuartersUseCase
 import com.gdavidpb.tuindice.record.domain.usecase.RemoveQuarterUseCase
