+++ summary/src/main/kotlin/com/gdavidpb/tuindice/summary/presentation/reducer/SummaryReducer.kt
@@ -3,7 +3,7 @@ package com.gdavidpb.tuindice.summary.presentation.reducer
 import com.gdavidpb.tuindice.base.domain.model.Account
 import com.gdavidpb.tuindice.base.domain.usecase.base.UseCaseState
 import com.gdavidpb.tuindice.base.presentation.reducer.BaseReducer
-import com.gdavidpb.tuindice.base.presentation.reducer.ViewOutput
+import com.gdavidpb.tuindice.base.utils.extension.ViewOutput
 import com.gdavidpb.tuindice.base.utils.ResourceResolver
 import com.gdavidpb.tuindice.summary.domain.usecase.error.GetAccountError
 import com.gdavidpb.tuindice.summary.presentation.contract.Summary
